python-client: update to latest version (#3)
The latest `python-client` version has some breaking changes in various
modules, requiring a small refactor. This PR updates the
`whisper-s2t-blockifier` to use the latest `python-client` as part of a
larger migration effort for all plugins and packages.

It is being sent to the `staging` branch. Moving to `main` will require
updating the `steamship.json` file to use the `prod` handle, and
updating the jupyter notebook invocation to match.